backend/requirements.txt
django-cors-headers
pylint_django
google-generativeai
python-dotenv
channels